fix a test failure under disable_tracing build tag
